+++ hack/tests/e2e-ansible.sh
@@ -92,8 +92,25 @@ test_operator() {
         exit 1
     fi
 
+    header_text "Verify that a config map owned by the CR has been created."
+    if ! timeout 1m bash -c -- "until kubectl get configmap test-blacklist-watches > /dev/null 2>&1; do sleep 1; done";
+    then
+        error_text "FAIL: Unable to retrieve config map test-blacklist-watches."
+        operator_logs
+        exit 1
+    fi
+
+    header_text "Verify that config map requests skip the cache."
+    if ! kubectl logs deployment/memcached-operator -c operator | grep -e "Skipping cache lookup\".*"Path\":\"\/api\/v1\/namespaces\/default\/configmaps\/test-blacklist-watches\";
+    then
+        error_text "FAIL: test-blacklist-watches should not be accessible with the cache."
+        operator_logs
+        exit 1
+    fi
+
+
     header_text "verify that metrics reflect cr creation"
-    if ! timeout 1m bash -c -- "until kubectl run -it --rm --restart=Never test-metrics --image=$metrics_test_image -- curl http://memcached-operator-metrics:8686/metrics | grep example-memcached; do sleep 1; done";
+    if ! timeout 60s bash -c -- "until kubectl run -it --rm --restart=Never test-metrics --image=$metrics_test_image -- curl http://memcached-operator-metrics:8686/metrics | grep example-memcached; do sleep 1; done";
     then
         error_text "FAIL: Failed to verify custom resource metrics"
         operator_logs