+++ internal/scorecard/plugins/basic_tests.go
@@ -21,6 +21,7 @@ import (
 	"strings"
 
 	schelpers "github.com/operator-framework/operator-sdk/internal/scorecard/helpers"
+	scapiv1alpha2 "github.com/operator-framework/operator-sdk/pkg/apis/scorecard/v1alpha2"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -50,8 +51,7 @@ func NewCheckSpecTest(conf BasicTestConfig) *CheckSpecTest {
 		TestInfo: schelpers.TestInfo{
 			Name:        "Spec Block Exists",
 			Description: "Custom Resource has a Spec Block",
-			Cumulative:  false,
-			Labels: map[string]string{necessityKey: requiredNecessity, suiteKey: basicSuiteName,
+			Labels:      map[string]string{necessityKey: requiredNecessity, suiteKey: basicSuiteName,
 				testKey: getStructShortName(CheckSpecTest{})},
 		},
 	}
@@ -70,8 +70,7 @@ func NewCheckStatusTest(conf BasicTestConfig) *CheckStatusTest {
 		TestInfo: schelpers.TestInfo{
 			Name:        "Status Block Exists",
 			Description: "Custom Resource has a Status Block",
-			Cumulative:  false,
-			Labels: map[string]string{necessityKey: requiredNecessity, suiteKey: basicSuiteName,
+			Labels:      map[string]string{necessityKey: requiredNecessity, suiteKey: basicSuiteName,
 				testKey: getStructShortName(CheckStatusTest{})},
 		},
 	}
@@ -106,9 +105,9 @@ func NewBasicTestSuite(conf BasicTestConfig) *schelpers.TestSuite {
 		"Test suite that runs basic, functional operator tests",
 	)
 
-	ts.AddTest(NewCheckSpecTest(conf), 1.5)
-	ts.AddTest(NewCheckStatusTest(conf), 1)
-	ts.AddTest(NewWritingIntoCRsHasEffectTest(conf), 1)
+	ts.AddTest(NewCheckSpecTest(conf))
+	ts.AddTest(NewCheckStatusTest(conf))
+	ts.AddTest(NewWritingIntoCRsHasEffectTest(conf))
 
 	return ts
 }
@@ -117,45 +116,52 @@ func NewBasicTestSuite(conf BasicTestConfig) *schelpers.TestSuite {
 
 // Run - implements Test interface
 func (t *CheckSpecTest) Run(ctx context.Context) *schelpers.TestResult {
-	res := &schelpers.TestResult{Test: t, MaximumPoints: 1}
+	res := &schelpers.TestResult{Test: t, State: scapiv1alpha2.PassState}
+
 	err := t.Client.Get(ctx, types.NamespacedName{Namespace: t.CR.GetNamespace(), Name: t.CR.GetName()}, t.CR)
 	if err != nil {
 		res.Errors = append(res.Errors, fmt.Errorf("error getting custom resource: %v", err))
+		res.State = scapiv1alpha2.ErrorState
 		return res
 	}
 
 	if t.CR.Object["spec"] == nil {
 		res.Suggestions = append(res.Suggestions, "Add a 'spec' field to your Custom Resource")
+		res.State = scapiv1alpha2.FailState
 		return res
 	}
-	res.EarnedPoints++
 	return res
 }
 
 // Run - implements Test interface
 func (t *CheckStatusTest) Run(ctx context.Context) *schelpers.TestResult {
-	res := &schelpers.TestResult{Test: t, MaximumPoints: 1}
+	res := &schelpers.TestResult{Test: t, State: scapiv1alpha2.PassState}
+
 	err := t.Client.Get(ctx, types.NamespacedName{Namespace: t.CR.GetNamespace(), Name: t.CR.GetName()}, t.CR)
 	if err != nil {
 		res.Errors = append(res.Errors, fmt.Errorf("error getting custom resource: %v", err))
+		res.State = scapiv1alpha2.ErrorState
 		return res
 	}
 	if t.CR.Object["status"] == nil {
 		res.Suggestions = append(res.Suggestions, "Add a 'status' field to your Custom Resource")
 		return res
 	}
-	res.EarnedPoints++
 	return res
 }
 
 // Run - implements Test interface
 func (t *WritingIntoCRsHasEffectTest) Run(ctx context.Context) *schelpers.TestResult {
-	res := &schelpers.TestResult{Test: t, MaximumPoints: 1}
+	res := &schelpers.TestResult{Test: t, State: scapiv1alpha2.PassState}
+
 	logs, err := getProxyLogs(t.ProxyPod)
 	if err != nil {
 		res.Errors = append(res.Errors, fmt.Errorf("error getting proxy logs: %v", err))
+		res.State = scapiv1alpha2.FailState
 		return res
 	}
+
+	var writes bool
 	msgMap := make(map[string]interface{})
 	for _, msg := range strings.Split(logs, "\n") {
 		if err := json.Unmarshal([]byte(msg), &msgMap); err != nil {
@@ -166,14 +172,15 @@ func (t *WritingIntoCRsHasEffectTest) Run(ctx context.Context) *schelpers.TestRe
 			continue
 		}
 		if method == "PUT" || method == "POST" {
-			res.EarnedPoints = 1
+			writes = true
 			break
 		}
 	}
 
-	if res.EarnedPoints != 1 {
-		res.Suggestions = append(res.Suggestions, "The operator should write into objects to update state."+
-			" No PUT or POST requests from the operator were recorded by the scorecard.")
+	if !writes {
+		res.Suggestions = append(res.Suggestions, "The operator should write into objects to update state."+ 
+		 	"No PUT or POST requests from the operator were recorded by the scorecard.")
+		res.State = scapiv1alpha2.FailState
 	}
 	return res
 }