+++ internal/scorecard/plugins/olm_tests.go
@@ -25,7 +25,7 @@ import (
 	"github.com/operator-framework/api/pkg/manifests"
 	schelpers "github.com/operator-framework/operator-sdk/internal/scorecard/helpers"
 	"github.com/operator-framework/operator-sdk/internal/util/k8sutil"
-	scapiv1alpha1 "github.com/operator-framework/operator-sdk/pkg/apis/scorecard/v1alpha1"
+	scapiv1alpha2 "github.com/operator-framework/operator-sdk/pkg/apis/scorecard/v1alpha2"
 	"github.com/sirupsen/logrus"
 
 	olmapiv1alpha1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/apis/operators/v1alpha1"
@@ -67,8 +67,8 @@ func NewBundleValidationTest(conf OLMTestConfig) *BundleValidationTest {
 		TestInfo: schelpers.TestInfo{
 			Name:        "Bundle Validation Test",
 			Description: "Validates bundle contents",
-			Cumulative:  false,
-			Labels: map[string]string{necessityKey: requiredNecessity, suiteKey: olmSuiteName,
+			Cumulative:  true,
+			Labels:      map[string]string{necessityKey: requiredNecessity, suiteKey: olmSuiteName, 
 				testKey: getStructShortName(BundleValidationTest{})},
 		},
 	}
@@ -175,11 +175,11 @@ func NewOLMTestSuite(conf OLMTestConfig) *schelpers.TestSuite {
 		"Test suite checks if an operator's CSV follows best practices",
 	)
 
-	ts.AddTest(NewBundleValidationTest(conf), 1)
-	ts.AddTest(NewCRDsHaveValidationTest(conf), 1.25)
-	ts.AddTest(NewCRDsHaveResourcesTest(conf), 1)
-	ts.AddTest(NewSpecDescriptorsTest(conf), 1)
-	ts.AddTest(NewStatusDescriptorsTest(conf), 1)
+	ts.AddTest(NewBundleValidationTest(conf))
+	ts.AddTest(NewCRDsHaveValidationTest(conf))
+	ts.AddTest(NewCRDsHaveResourcesTest(conf))
+	ts.AddTest(NewSpecDescriptorsTest(conf))
+	ts.AddTest(NewStatusDescriptorsTest(conf))
 
 	return ts
 }
@@ -188,11 +188,12 @@ func NewOLMTestSuite(conf OLMTestConfig) *schelpers.TestSuite {
 
 // Run - implements Test interface
 func (t *BundleValidationTest) Run(ctx context.Context) *schelpers.TestResult {
-	res := &schelpers.TestResult{Test: t, MaximumPoints: 1}
+	res := &schelpers.TestResult{Test: t, State: scapiv1alpha2.PassState}
 
 	if t.OLMTestConfig.Bundle == "" {
-		res.Errors = append(res.Errors,
+		res.Errors = append(res.Errors, 
 			errors.New("unable to find the OLM 'bundle' directory which is required for this test"))
+		res.State = scapiv1alpha2.ErrorState
 		return res
 	}
 
@@ -207,6 +208,7 @@ func (t *BundleValidationTest) Run(ctx context.Context) *schelpers.TestResult {
 	for _, result := range validationResults {
 		for _, e := range result.Errors {
 			res.Errors = append(res.Errors, &e)
+			res.State = scapiv1alpha2.FailState
 		}
 
 		for _, w := range result.Warnings {
@@ -215,10 +217,6 @@ func (t *BundleValidationTest) Run(ctx context.Context) *schelpers.TestResult {
 	}
 
 	res.Log = validationLogOutput.String()
-
-	if len(res.Errors) == 0 {
-		res.EarnedPoints++
-	}
 
 	return res
 }
@@ -239,17 +237,17 @@ func matchVersion(version string, crd *apiextv1beta1.CustomResourceDefinition) b
 
 // Run - implements Test interface
 func (t *CRDsHaveValidationTest) Run(ctx context.Context) *schelpers.TestResult {
-	res := &schelpers.TestResult{Test: t}
+	res := &schelpers.TestResult{Test: t, State: scapiv1alpha2.PassState}
 	crds, err := k8sutil.GetCRDs(t.CRDsDir)
 	if err != nil {
 		res.Errors = append(res.Errors, fmt.Errorf("failed to get CRDs in %s directory: %v", t.CRDsDir, err))
-		res.State = scapiv1alpha1.ErrorState
+		res.State = scapiv1alpha2.ErrorState
 		return res
 	}
 	err = t.Client.Get(ctx, types.NamespacedName{Namespace: t.CR.GetNamespace(), Name: t.CR.GetName()}, t.CR)
 	if err != nil {
 		res.Errors = append(res.Errors, err)
-		res.State = scapiv1alpha1.ErrorState
+		res.State = scapiv1alpha2.ErrorState
 		return res
 	}
 	for _, crd := range crds {
@@ -259,7 +257,6 @@ func (t *CRDsHaveValidationTest) Run(ctx context.Context) *schelpers.TestResult 
 		if !(matchVersion(gvk.Version, crd) && matchKind(gvk.Kind, crd.Spec.Names.Kind)) {
 			continue
 		}
-		res.MaximumPoints++
 		if crd.Spec.Validation == nil {
 			res.Suggestions = append(res.Suggestions, fmt.Sprintf("Add CRD validation for %s/%s",
 				crd.Spec.Names.Kind, crd.Spec.Version))
@@ -287,8 +284,9 @@ func (t *CRDsHaveValidationTest) Run(ctx context.Context) *schelpers.TestResult 
 				}
 			}
 		}
-		if !failed {
-			res.EarnedPoints++
+
+		if failed {
+			res.State = scapiv1alpha2.FailState
 		}
 	}
 	return res
@@ -296,15 +294,12 @@ func (t *CRDsHaveValidationTest) Run(ctx context.Context) *schelpers.TestResult 
 
 // Run - implements Test interface
 func (t *CRDsHaveResourcesTest) Run(ctx context.Context) *schelpers.TestResult {
-	res := &schelpers.TestResult{Test: t}
+	res := &schelpers.TestResult{Test: t, State: scapiv1alpha2.PassState}
+
 	var missingResources []string
 	for _, crd := range t.CSV.Spec.CustomResourceDefinitions.Owned {
 		gvk := t.CR.GroupVersionKind()
 		if strings.EqualFold(crd.Version, gvk.Version) && matchKind(gvk.Kind, crd.Kind) {
-			res.MaximumPoints++
-			if len(crd.Resources) > 0 {
-				res.EarnedPoints++
-			}
 			resources, err := getUsedResources(t.ProxyPod)
 			if err != nil {
 				log.Warningf("getUsedResource failed: %v", err)
@@ -329,7 +324,9 @@ func (t *CRDsHaveResourcesTest) Run(ctx context.Context) *schelpers.TestResult {
 		res.Suggestions = append(res.Suggestions, fmt.Sprintf("If it would be helpful to an end-user to"+
 			" understand or troubleshoot your CR, consider adding resources %v to the resources section for owned"+
 			" CRD %s", missingResources, t.CR.GroupVersionKind().Kind))
+		res.State = scapiv1alpha2.FailState
 	}
+
 	return res
 }
 
@@ -435,11 +432,12 @@ func getUsedResources(proxyPod *v1.Pod) ([]schema.GroupVersionKind, error) {
 
 // Run - implements Test interface
 func (t *StatusDescriptorsTest) Run(ctx context.Context) *schelpers.TestResult {
-	res := &schelpers.TestResult{Test: t}
+	res := &schelpers.TestResult{Test: t, State: scapiv1alpha2.PassState}
+
 	err := t.Client.Get(ctx, types.NamespacedName{Namespace: t.CR.GetNamespace(), Name: t.CR.GetName()}, t.CR)
 	if err != nil {
 		res.Errors = append(res.Errors, err)
-		res.State = scapiv1alpha1.ErrorState
+		res.State = scapiv1alpha2.ErrorState
 		return res
 	}
 
@@ -448,11 +446,11 @@ func (t *StatusDescriptorsTest) Run(ctx context.Context) *schelpers.TestResult {
 
 // Run - implements Test interface
 func (t *SpecDescriptorsTest) Run(ctx context.Context) *schelpers.TestResult {
-	res := &schelpers.TestResult{Test: t}
+	res := &schelpers.TestResult{Test: t, State: scapiv1alpha2.PassState}
 	err := t.Client.Get(ctx, types.NamespacedName{Namespace: t.CR.GetNamespace(), Name: t.CR.GetName()}, t.CR)
 	if err != nil {
 		res.Errors = append(res.Errors, err)
-		res.State = scapiv1alpha1.ErrorState
+		res.State = scapiv1alpha2.ErrorState
 		return res
 	}
 
@@ -462,10 +460,11 @@ func (t *SpecDescriptorsTest) Run(ctx context.Context) *schelpers.TestResult {
 func checkOwnedCSVDescriptors(cr *unstructured.Unstructured, csv *olmapiv1alpha1.ClusterServiceVersion,
 	descriptor string, res *schelpers.TestResult) *schelpers.TestResult {
 	if cr.Object[descriptor] == nil {
+		res.State = scapiv1alpha2.FailState
 		return res
 	}
 	block := cr.Object[descriptor].(map[string]interface{})
-	res.MaximumPoints = len(block)
+
 	var crd *olmapiv1alpha1.CRDDescription
 	for _, owned := range csv.Spec.CustomResourceDefinitions.Owned {
 		if owned.Kind == cr.GetKind() {
@@ -475,6 +474,7 @@ func checkOwnedCSVDescriptors(cr *unstructured.Unstructured, csv *olmapiv1alpha1
 	}
 
 	if crd == nil {
+		res.State = scapiv1alpha2.FailState
 		return res
 	}
 
@@ -482,7 +482,6 @@ func checkOwnedCSVDescriptors(cr *unstructured.Unstructured, csv *olmapiv1alpha1
 		for key := range block {
 			for _, statDesc := range crd.StatusDescriptors {
 				if statDesc.Path == key {
-					res.EarnedPoints++
 					delete(block, key)
 					break
 				}
@@ -493,7 +492,6 @@ func checkOwnedCSVDescriptors(cr *unstructured.Unstructured, csv *olmapiv1alpha1
 		for key := range block {
 			for _, specDesc := range crd.SpecDescriptors {
 				if specDesc.Path == key {
-					res.EarnedPoints++
 					delete(block, key)
 					break
 				}
@@ -503,6 +501,7 @@ func checkOwnedCSVDescriptors(cr *unstructured.Unstructured, csv *olmapiv1alpha1
 
 	for key := range block {
 		res.Suggestions = append(res.Suggestions, fmt.Sprintf("Add a %s descriptor for %s", descriptor, key))
+		res.State = scapiv1alpha2.FailState
 	}
 	return res
 }